Changed  ESD chip, changed Type C routing
ESD I wanted to order was not in stock, so I changed to a different part in the schematic.

Upon reading the datasheet, it was made clear the +5V line has t pass through the ESD chip from one side to another, not merely run a trace into and out of it from one side.