Added the logo
Added the logo to the symbol and footprint library, then added it to the schematic and then PCB.

If the logo has to be edited, this should be done by replacing the footprint for it.

Next step is adding a revision/version number to the logo itself.
Maybe some silly 'made by' type of thing too.